[LevelZero] Add LevelZero to HAL and Codegen
Use flags -DIREE_BUILD_EXPERIMENTAL_LEVEL_ZERO=ON -DLEVEL_ZERO_HEADERS_API_ROOT=/home/stanley/nod/level-zero
-LevelZero HAL Driver
-Addi OpenCL HAL Target compiler
-Add SPIRV Codegen for Kernel capability
-fix illegal pointer arithmetic on void* by Boian
-Use events for command buffer execution and synchronization (#47)
-Add flag for switching between Physical64 and Physical32 addressing in OpenCL
-enable creation of device by UUID + ID(device handle as uintptr_t)
Note that the device ID implemented like that is ephemeral and is valid
only in the current IREE runtime context. If you start a new process
the IDs will be different.
With this change you can do

$ iree-run-module --list_devices
level_zero://00005100-0000-0000-0000-000000000001

$ iree-run-module --device=level_zero://00005100-0000-0000-0000-000000000001 ...

-add query_memory_heaps implementation

Fixes
error: arithmetic on a pointer to void is a GNU extension [-Werror,-Wgnu-pointer-arith]

-Supply structure type when passing such arguments in accordance with the API (#34)

When calling Level Zero API functions that query information and use a struct to
populate the information, the user must supply the structure type (stype) in
the structure itself. The struct is both in and out argument.