Add Winograd support for NCHW convolutions
This patch adds support for NCHW convolutions.
In order to avoid the cost of additional transposes,
the winograd input op is modified so that regardless
of whether the input is (N, H, W, C) or (N, C, H, W),
the output is always (T, T, N, H, W, C). Similarly,
the winograd output op is modified to have the input
always be (T, T, N, H, W, C), but have the output be
either (N, H, W, C) or (N, C, H, W) depending on the
specified image dimensions. The tiling and decomposition
passes are also modified to account for this.