[Preprocessing] Add pass to fuse dequantization and matmul linalg.generic ops
This pass finds a linalg.GenericOp representing a matmul and places it in a
dispatch region with its preceding op if the preceding op is a linalg.GenericOp
representing a dequantization. Fusing these ops prevents loading the result of
the dequantization into and out of memory. Currently, the resulting dispatch after
this fusion will be assigned an appropriate tiling configuration.