Only look for VectorX if it's part of layout otherwise assume step=1.
Currently, WMMA op has a [vectorY, laneY] = [8, 2] for cMLayout. This is
assuming vectorX is == 1. However, our distirbution pattern does not
expect that. This patch teaches the pattern to expect it.

This seems to be OK for distributing just contraction, however, during
distribution of MultiDimReduction, when we compute SIMTIndex, it ends up
generating really large number as indices/step. This patch solves that
issue.